pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
]
